+++ src/main/java/com/sanitas/calculator/domain/operations/OperationFactory.java
@@ -1,5 +1,6 @@
 package com.sanitas.calculator.domain.operations;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.sanitas.calculator.domain.exception.OperationException;
@@ -11,5 +12,5 @@ import com.sanitas.calculator.domain.exception.OperationException;
 public interface OperationFactory {
 	String ARGUMENT_QUANTITY_ERROR_TEXT = "Number of arguments not valid.";
 
-	Double calculate(List<Double> arguments) throws OperationException;
+	BigDecimal calculate(List<BigDecimal> arguments) throws OperationException;
 }
